Bugfix: Catch JUnit-specific ArgumentsAreDifferent-Exception
If Mockito is run from JUnit, instead of a normal ArgumentsAreDifferentException, a different ArgumentsAreDifferent from org.mockito.exceptions.verification.junit was thrown. This caused the timeout-loop to exit instantly instead of waiting for the timeout to pass.

This fixes the problem by checking for the JUnit-specific exception and converting it to a MockitoAssertionError.