Comparison Should Use Equals Method Instead of ==
Hi, 

I am writing framework that relies on Mockito which requires proxying mocks when running concurrent test environment. The comparison operator in MockHandlerImpl does not give me the opportunity to intercept equality check during verification. If equals() is used instead of == operator it would give me the opportunity to intercept calls to equals() and to do a proper comparison with my proxy's underlying mock instance.