Fixes #200 : Tweak generic signature of ArgumentCaptor.fromClass
We modify the signature of ArgumentCaptor.fromClass(Class<T>) to better express the type relationship between the argument and the return type.  There is no change to the behaviour of the method.  This change is to allow expressions such as:

ArgumentCaptor<Consumer<String>> captor = ArgumentCaptor.fromClass(Consumer.class)

to type check, which is desirable as a convenience to users of ArgumentCaptor.

We also add two tests to document this extension to the api, both of which fail to compile without this change to ArgumentCaptor.fromClass.  These tests are unusual in that they make no assertions; this is justified because the change which they test is a change to the formal, or compile-time, properties of ArgumentCaptor, not its behaviour.