Change processing of combination of @Spy and @Injectmocks. Fixes #285
MockUtil can not determine the name of a field annotated with
both @Spy and @Injectmocks which made the NameBasedFilter fail.
Such an object internally was not a mock, but a real object
and therefore it does not have a MockHandler.

However, as this is a real object, we can treat it as-is. Therefore if
a field is annotated with @Spy and @Injectmocks, we treat it as @Real
instead of a @Spy.