Fully register a class' ancestry for GenericMetadataSupport.
The old implemention would take a class X and consider all of its superclasses
and the interfaces implemented by X and the interfaces implemented by all of
its superclasses.

The new implementation also considers the superinterfaces of the interfaces
implemented by X and the superinterfaces of the interfaces implemented by its
superclasses.

Fixes #497