Strict rules - fail fast when arg mismatches
Initial, prototype-style implementation of the new feature of the strict rules: failing immediately when incorrect argument used on the stubbed method.

This feature may trigger false negative signals. It's a trade-off. In most of the cases, wrong argument is a problem in test/code that needs to be debugged. For the remaining scenarios, users can opt-out from strictness (or declare optional stubbings, API for that will be added later, not part of this PR).

Work in progress but review feedback welcome.