Add missing newlines at the end of source files
At the time this patch was authored, 252 (~31%) source files in the
project did not end with a newline, while 555 (~69%) did.

This patch takes the de-facto standard of having a newline at the end
of a file and applies it to all the source files in the project.

Besides improving standardization, another contributing consideration
is the fact that many editors/IDEs add newlines at the end of files.
Using such an IDE will add a newline to the edited file, adding
"noise" to the patch and confusing potential reviewers. By explicitly
adding all the missing newlines in a single patch, this problem can be
avoided.

These changes were done automatically by applying the following shell
command:

$ find . -name "*.java" -exec sed -i -e '$a\' {} \;