Fixes #992: Add MockitoV1CompatibilityMatcher to ease migration from pre 2.x
The 1.x version of ArgumentMatcher<T> differs quite
significantly from the current version:
1) It is a class not an interface and is not standalone as it
   extends org.hamcrest.BaseMatcher.
2) Its main method is matches(Object) not matches(T).
3) It uses the describeTo(Description) method to construct a
   description and not the toString() method.

This class supports the following strategy for migrating
implementations of ArgumentMatcher from 1.10.19 to 2.x:
1) First migrate to a 1.x version that has
   MockitoV1CompatibilityMatcher change in it. That should be
   relatively simple.
2) Then modify all existing implementation of ArgumentMatcher
   to extend MockitoV1CompatibilityMatcher.
3) Switch to use a version of 2.x that contains this change.
   Everything should compile as usual.
4) Modify the classes that extend MockitoV1CompatibilityMatcher
   to implement ArgumentMatcher interface directly.

The above strategy is overkill when migrating a small project
that can be updated atomically but it has proven to be a good
approach when upgrading a large project spread across multiple
git repositories. This is added in part because of the
experience doing exactly that with the Android Open Source
Platform.

Duplicates some of the code from TypeSafeMatching to allow
MockitoV1CompatibilityMatcher to provide the same type safe
matching capabilities as is provided for direct implementations
of ArgumentMatcher.