Fixes #976: Resolve ambiguous constructors (#980)
* Fixes #976: Improve constructor choosing

With the current code (introduced in Mockito 2.7.14 by commit
6a82c03), calling MockSettings#useConstructor with an argument list
that would be applicable to more than one constructor would fail with
an org.mockito.internal.creation.instance.InstantiationException.

This behavior, however, is suboptimal, as described in issue #976, as
it makes useConstructor less robust than the Java compiler, which is
able to resolve such ambiguities.
With this patch, Mockito will attempt to match the constructor with
the most specific parameter types. A constructor X is considered more
specific than a constructor Y if:
1. They are both applicable to the given argument list
2. Constructor X has at least one parameter which is a further
   specialization of the corresponding parameter of constructor Y
   (i.e. paramX.isAssignableFrom(paramY)).
3. Constructor Y has no parameter which is a further specialization
   of the corresponding parameter of constructor X, as defined above.

E.g., consider the following class:

    public class SomeClass {
        SomeClass(Object o) {}
        SomeClass(String s) {}
    }

Without this patch, calling

    mock(SomeClass.class, withSettings().useConstructor("string!"))

would fail. With this patch, such a call would invoke the
SomeClass(String) constructor.

* Fix Freudian slip in message regarding issue 685

The text should read "a better JOB", not "a better JOIN".