Fix several javadoc issues (#1040)
* SilentRunnerTest#HasUnnecessaryStubs javadoc

This patch moves the class' javadoc before the annotation to solve a
"dangling javadoc" warning.

* ArgumentMatchingToolTest#shouldUseMatchersSafely comment

The inner StringMatcher is not a top level or member class, it's a
local variable.
As such, using a javadoc style comment (/** */) on it generates a
"dangling javadoc" warning. This patch changes that javadoc comment
to a regular java comment (//) to resolve that warning.

* @see tag references to bugs

If a @see tag doesn't reference another class/method, the link should
be manually handled by a <a> HTML tag.

* VarargsTest#shouldCaptureVarArgsAsArray @throws tag

Remove a @throws tag for the checked Exception that isn't thrown by
this method.