Fixes #1050 : More specific error if can't mock due to ClassNotFound
If a ClassNotFoundException is the root cause of not being able to mock
an object, present that information to the user, instead of saying it's
probably because the class is private or final. This can avoid a lot of
confusion when a user is trying to figure out why they can't mock an
object.