Fixes #1313: Use weak refs to mock in invocations
This is needed for mock makers that store a
map<weakref<mock>, invocation>. If the invocation holds a strong
reference to the mock, the key never gets gc-ed and the map keeps on
growing.

As weak reference cannot be serialized, temporarily create a strong
reference during serialization.

As the tests use invocations without keeping references to the mock we
need to keep strong references in these cases.