Fixed bug in MockitoSession API
When initMocks() fails we did not clean up the session listener. This led to a confusing error message on the next attempt to create session. I stumbled upon this problem when attempting to configure TestNG to use strict stubbing.

In order to fix this cleanly I need to keep the existing functionality that protects the user from forgetting to use 'finishMocking()'. Hence, I still need to throw an exception if the user adds the same listener multiple times. However, in the event that the listener is dirty, I clean it up automatically instead of failing.