Prevent StackOverflowError when processing reversed generics (#1565)
When processing the reverse method, the type parameters are actually
reversed compared to the type parameters in the class. This subsequently
confuses GenericMetadataSupport, as it creates a cycle in
`contextualActualTypeParameters`. In that case A resolves to B, while B
resolves to A.

The root cause is that we are processing type parameters of methods,
while we already processed them in the class declaration. Therefore,
ignore any type parameters if we already derived the appropriate
contextual type parameter.