Return null instead of causing a CCE (#1612)
This solves a large number of edge-cases where `null` will actually
remove the runtime ClassCastException. This essentially negates the
whole MockitoCast ErrorProne check. We can still not support every use
case, but causing a NPE instead of a CCE does not seem to make this
worse.

I am still running internal tests within Google to see if there are any
regressions, but I already saw that some of the test failures we had
with ByteBuddy were resolved with this particular patch.

Note that this now fully closes #357. A previous PR resolved the same
issue with ReturnsSmartNulls: #1576.

Fixes #357