Enable spotless to automatically format our source code
Spotless [1] has been used by various popular open source projects,
including JUnit [2]. While working on Mockito inside of Google, I
discovered that our formatting is inconsistent. We can use Spotless to
automatically take care of that.

To be able to format the source code, run

./gradlew spotlessApply

This check runs on Travis to make sure code remains formatted. This
should also reduce the amount of trouble with the CheckStyle linter, as
Spotless will automatically resolve these issues.

On purpose, the spotless configuration is quite minimal. The current
configuration is mostly a reflection on the current state of affairs. We
can later discuss different configuration options and how we can
integrate them.

[1]: https://github.com/diffplug/spotless/tree/master/plugin-gradle
[2]: https://github.com/junit-team/junit5/blob/3f491f3148d2c745808fc75b0a802b60243e104e/build.gradle.kts#L124-L144