Add default implementation for VerificationMode#description
This method was added in Mockito 2 [1] and was implemented in all
relevant subclasses. However, most subclasses do not implement a custom
description.

Since we are now building on Java 8, we can ship this as a default
method and remove all duplicate subclass implementations. As a
side-effect, it allows users of Mockito 1 to migrate to Mockito 3
without additional breakages. (They could implement it, but
VerificationModeFactory is internal and would require additional
boilerplate)

[1]: https://github.com/mockito/mockito/pull/68