Fixes #1839 : Badge improvements in README.md (#1840)
Use shields.io badge for Bintray, as the Bintray hosted badge shows
"Download" and cannot be made to show "Bintray".

Split an excessively long line for the first row of badges.

Update the alt text for "Release Notes". Non-graphical users would only
see a link to the release notes.

"Javadoc" was misspelled as "Javadocs".

Use HTTPS for all links.

Use direct links to the package on Maven Central and Bintray.

Use the default color for Javadoc, red looks as an error.