Add TestRule to MockitoJUnit
JUnit has two types of test rules: MethodRule and TestRule. The existing
MockitoJUnit rule returns a MockitoRule which extends MethodRule.
However, since this is not a TestRule, some features of JUnit do not
nicely integrate with Mockito. For example, the RuleChain feature of
JUnit only works on TestRules.

Therefore, add a separate entrypoint to MockitoJUnit which returns a
TestRule that can be integrated with RuleChain. It introduces a separate
interface to make this distinction clear. If we would change the
existing MockitoRule to also extend TestRule, the JUnit logic would no
longer run the MethodRule [1] and would thus break existing users.

[1]: https://github.com/junit-team/junit4/blob/2df7e0882128d551565f87f688bbe745d85aacba/src/main/java/org/junit/runners/BlockJUnit4ClassRunner.java#L402