Fixes #1891: Instruct Gradle to build JARs deterministically (#1892)
This makes the build reproducible, if there are no other sources of non-determinism with the JAR's content such as the manifest.

It strips the "Bnd-LastModified" header from OSGi content in JARs. it copies the jar and removes the offending header from the Manifest file. Due to a known issue in Gradle with the (now-deprecated) OSGi plugin, this is one of the only ways to remove this header.

Additionally, we verify the reproducibility in CI build via a new Gradle task. This new task, `checkReproducibility`, runs two builds, collects the sha256 checksums of the output archives, and compares them.

Co-authored-by: Tim van der Lippe <tvanderlippe@google.com>