Fixes #1905 : Fix Pattern matcher not matching to subregion (#1914)
Issue #1905 talked about a particular pattern using Java's Pattern that
causes Mockito to not properly match the argument (provided with an
invocation) to the expected answer. It was later discovered that the
Pattern matcher matches to the entire region, not some subregion, by
using Java's Matcher.match(). This commit fixes that by using
Matcher.find().