Move spotless check to separate build task
This means the tests can run in parallel of the formatting changes.
Therefore, formatting changes would not block obtaining your test
results, which should hopefully reduce the amount of Travis builds
necessary to work on a community PR.

This is also in preparation of #1934 which requires spotless to
run on JDK11 minimum.