Support mocking of function with Kotlin inline classes (#2280)
Despite functions with inline class parameters being already supported,
there were not tests for it. Thus, add them.

The other part is functions, returning inline classes. There can be several
kinds of inline classes - with primitive underlying type (nullable or not-null)
and reference types (nullable or not-null). Also, a function can return either
nullable of not-null inline class. If a function returns nullable inline class,
we do not do anything special - the function's signature contains inline class.
However, if the function returns not-null inline class, its signature contains
underlying type, thus, we need to unbox the inline class (i.e. get underlying
type's value) to mock the return value.

For example, if we have the following inline class

  inline class IC(val i: Int)

and we want to mock return value of a function, returning the inline class

  fun returnsIC(): IC

like

  `when`(mock.returnsIC()).thenReturn(IC(1))

then if we do not unbox the argument of `thenReturn` function, we get error
about incompatible types.

The only issue, however, is that we cannot check, whether the function indeed
returns `IC`, unless we use kotlin.reflect. So, the following code will work.

  fun returnsInt(): Int

  `when`(mock.returnsInt()).thenReturn(IC(1))

Fixes issue #1930