Remove all Incubating annotations (#2437)
For all of these APIs, we have been shipping them for at least 1 major
release. We consider all of these APIs part of the public API contract
now and therefore we should mark them as such.

This PR removes all usages of the annotation. If we were to later remove
an API, it will first have to be deprecated and then removed in a
subsequent major release. However, we have no plans to do so with any of
these APIs.