Issue-2544 Fixed bug when more than 2 invocations in Mockito.mockConstructionWithAnswer().
If you build a mockConstructionWithAnswer with more than 1 additionalAnswers there was a logic error
whereby the method would return the wrong value when making the second last invocation. It would
accidentally think it's at the end an only return the last answer, never returning the second
last answer when it should.

The changes I made to ConstructionMockTest.java now exercises this condition.