CapturingMatcher
BREAKING CHANGE: The behaviour of argument captors when passed to a varargs parameter has changed. Prior to Mockito v5, passing an argument captor to a varargs parameter would capture any number of values passed to the varargs parameter. From Mockito v5, the same call will match only a single value passed to the varargs parameter.  To capture any number of values, change to the type of the argument captor to match the array type of the varargs parameter. For example, given a `String...` varargs parameter, change the argument captor from type `String` to type `String[]`.