Fix copyMockMethodDispatcher task
The task was incorrectly defining inputs and outputs, resulting in the task being marked 'UP-TO-DATE' and not re-running when the input changed or the output was removed.

The task now copies the 'raw' file into a new output directory and includes that directory in the jar.

This corrects the task inputs and outputs, while also avoiding warnings about shared output directories.