Catch NoClassDefFoundError in KotlinInlineClassUtil
When using R8 in AGP 7.4+, we noticed that this code block failed at runtime with `NoClassDefFoundError`. My theory for why this happens is that something in the build pipeline for Android is optimizing `Class.forName(...)` away into a linked `JvmInline.class` call, but then R8 strips the annotation as unused.

You may think - why would mockito be running under minified conditions??

We actually do this in Slack Android's instrumentation tests to create test conditions more similar to production environments, and have a gradle plugin called [Keeper](https://github.com/slackhq/keeper) that generates proguard rules for classes used in the main app _from_ androidTest sources or libraries.

I'm still investigating how this state occurs, but thought to send by this PR to see if you'd all be open to it as an extra catch.