Apply jacoco plugin to allprojects fully
The jacoco plugin was applied on the wrong Project object which lead to multiple workarounds:
 * somehow `apply plugin:` binds to the script default project which is rootProject
 * so the plugin wasn't applied to currentProject
 * so it didn't register .jacoco on the Test tasks
 * so manual `applyTo()` was required
 * calling `applyTo test` resolved to `rootProject.jacoco.applyTo(currentProject.tasks.test)` (because the currentProject.jacoco wasn't registered!)
 * which lead to a cross-project configuration leak, and we had the warning:

> Task :*:test
Resolution of the configuration :jacocoAgent was attempted from a context different than the project context. Have a look at the documentation to understand why this is a problem and how it can be resolved. This behavior has been deprecated. This will fail with an error in Gradle 9.0. For more information, please refer to https://docs.gradle.org/8.2/userguide/viewing_debugging_dependencies.html#sub:resolving-unsafe-configuration-resolution-errors in the Gradle documentation.