Added JavaFlightRecorder options for Tests
This allows you to create JFR recordings with the flag -Pjfr
when running Gradle Test tasks.

The jfr_config.jfc provides a profiling config for the JFR.
It tracks the TLAB allocations in addition to the normal
profiling config in JDKs (JDK/lib/jfr/profile.jfc).
It has also the following changes:
 * Method sampling interval of 2 ms
 * Thread dump  interval of 10 s
 * CompilerInlining true
 * ObjectAllocationSample throttle 1000/s
 * JavaExceptionThrow true