Resolve more Gradle Tasks lazily (#3099)
Changes the configuration and registration of Gradle Tasks to use
the lazy API of Gradle. This shall speed up the configuration phase
e.g. during single test execution.

Note: the tasks.withType() API has currently no overload with Action
as the tasks.named() has, so we need to configureEach to be lazy.