Attempt to detect system property mangling prior to loading ByteBuddy.
If the user stubs their System properties (e.g. to test code that handles System properties
to determine the platform), then they may unexpectedly hit the issue that ByteBuddy will
attempt to fork a subprocess upon being installed, and that relies on the JVM Process class.
Internally, the Process class makes a call to 'System.getProperty(os.name)', which results
in a NullPointerException being raised if the property does not exist.

When this exception occurs, it bubbles up to Mockito in the shape of a very confusing
error message, which we should ideally try to avoid.

This change adds a check for this edge case prior to loading ByteBuddy so that a meaningful
error message can be raised that instructs the user on how to mitigate this issue.