Clarify structure of commit messages (#3626)
Since I have been merging most of the community contributions,
I realized that I continuously rewrite commit messages with regards
to the `Fixes #id` notation.

Therefore, update the contribution guidelines to clarify that the
fixes part should come at the end (as to not clutter the message).
Also, link to the commit guide that I usually refer to, which also
clarifies that the commit title should start with a verb.