build: Allow overriding 'Created-By' for reproducible builds
In order to support bit-for-bit reproducible builds, this commit
introduces a Gradle property 'manifest.createdBy'. This property
allows rebuilders to override the 'Created-By' attribute in the
JAR manifest files.

If the property is not set, the build falls back to using the
current JVM's details, preserving the existing behavior. This directly
addresses the problem of non-reproducible builds due to different
JDK vendor information in the manifest.

Fixes #3563