Add support for generic types in `MockedConstruction` and `MockedStatic` (#3729)
MockedStatic and MockedConstruction threw a MockitoException when the
provided type was itself parameterized, suggesting to use a raw type instead.
For instance, MockedConstruction<MyGenericType<String>> was not allowed.

This commit allows to mock generic types without throwing an exception and
without forcing users to deal with 'raw use of parameterized class' warnings.

Fixes #2401