Feat: automatically detect class to mock in mockStatic and mockConstruction
As PR #2779 introduced the ability to automatically detect the class to mock
for mocks and spies, this commit allows doing the same with mockStatic
and mockConstruction.

Added overloaded reified methods to match all the counterparts that consume
an explicit Class literal parameter.

For instance, users can now leverage these:
* MockedStatic<SomeClass> mock = mockStatic();
* MockedConstruction<SomeClass> mock = mockConstruction();

instead of:
* MockedStatic<SomeClass> mock = mockStatic(SomeClass.class);
* MockedConstruction<SomeClass> mock = mockConstruction(SomeClass.class);

The following methods have been added:
* mockStatic(T...)
* mockStatic(Answer, T...)
* mockStatic(String, T...)
* mockStatic(MockSettings, T...)
* mockConstruction(T...)
* mockConstruction(MockInitializer, T...)
* mockConstruction(MockSettings, T...)
* mockConstruction(Function<Context, MockSettings>, T...)
* mockConstruction(MockSettings, MockInitializer, T...)
* mockConstruction(Function<Context, MockSettings>, MockInitializer, T...)